fix(*): make containers host-aware
Before, you had to hardcode the host's IP address and the IP address
to the etcd cluster. Now that we are on CoreOS, we can make some
assumptions about the system, including the default gateway and
where etcd is located.

The host IP address was always accessible through the default gateway,
so there's a change to use netstat to grep for the ip address
of the docker0 bridge. Using that, we can also assume that etcd
is clustered on the host vs. a remote container on some other node,
so we can just use the host IP address and a default port number
to retrieve the endpoint.

We should be using Fleet to manage these containers, so I removed
the dependencies on certain systemd units, since fleet already
considers those units to be running on each host. I also removed
the dependency on deis-builder to be running before deis-controller
can be started. We did this because we wanted to ensure that users
did not run into issues when running a git push, but I feel that it's
a poor decision to kill the controller if the builder is down. By that
argument, we should kill the controller when *any* component comes
down. A `fleetctl list-units` should be more than enough to ensure
that the containers are running as they should be.