fix(controller): update Django admin fields
The declarative definition for presenting Build model objects
in the Django admin still referenced "sha," which is no longer
a field on the Build model. Same for the "version" field on
Config. Both of these caused a 500 error when browsing at
http://local.deisapp.com:8000/admin/

I also added "created" to Build, Config, and Release and
"domain" to Cluster in the admin list display.