feat(controller): make Procfile mandatory
At the moment, all custom Dockerfile apps will not start because the
'start' command is not present in the container. Instead, we can
generically parse the Procfile for the process type and start that
instead. This makes two important changes:

1) All applications **must** specify a Procfile
2) All custom Dockerfile apps must have access to cat, grep, cut and sh

The justification for 1) is so that we have a known place to search for
runnable processes. 2) is because of the change from `start web` to parse
the Procfile for the process types.

fixes #668