fix(services): use COREOS_PUBLIC_IPV4 for advertised IP
Currently, all containers use a hacky method to determine the IP
to advertise to etcd - the method determines the machine's default
route, and uses that IP. However, this can be configured very differently
depending on the environment, and in some cases doesn't work at all.
Instead, we use COREOS_PUBLIC_IPV4, as that is set on every CoreOS host,
and we trust that CoreOS knows the machine's networks better than we do.

fixes #845