fix(controller): add source release version
When rolling back releases, the controller would publish a new
release, which meant that the app image would get a version
bump and a new tag on the registry would be created. However,
the tag that was created on the registry was based off the 'latest'
tag. This allows rollbacks to change the source tag to another
version so that the new tag will be based off the release requested.

fixes #1072