fix(tests): increase component test timeouts to 20min
Most component test failures in our CI system recently have been due
to the tests taking longer than the default Go testing package timeout
of 10 minutes. The tests are in progress, but Go kills the process
after 10 minutes and reports it as failed.

The tradeoff here is that if one of the tests is actually stuck,
we wait twice as long to find that out, but overall this seems like a
fix for CI stability. Builder was increased to 30m since it takes much
longer to pull slugbuilder.