fix(systemd units): Fix docker containers being orphaned
Docker seems to orphan our containers after the fix in 6b452e,
this reverts part of it and properly signals the docker containers to
stop. Given that we run them with `--rm`, this also makes the ephemeral
containers destroy their data off of the drives. The data containers are
not affected by this and they persist.