feat(controller): check client against API version
Before locking the API version, we needed some way to ensure that
we did not break the client. This was done by comparing the client's
version against the server. If there was a mismatch on the minor or
major release versions (as noted in Semver 2.0.0[1]), the server
would refuse the request. Because we have locked the API version,
it would make sense to test against that to see if the client is
compatible with the API. If not, then the client should upgrade.

This also modifies the current behaviour to warn the user on a minor
version mismatch. If there is a minor version mismatch between the
client and the server, it will press onwards and display a warning,
advising the user to upgrade. If there is a major version mismatch,
then the API server will refuse the request as before.

[1]: http://semver.org/