fix(controller): silence redundant app release logs
When logging events to the application, it does not need to know
which app it came from... That information is redundant. Also, the
developer does not need to know the internal UUIDs of the builds, config
and releases being created, as the relevant information will be in the
release summary.