ref(router): split domains to its own custom directive
In order to support multiple certs on multiple custom domains bound to
the same application, we need to split out each custom domain to its own
`server {}` directive. In master, the domain etcd keys are set like so:

    SET /deis/domains/go "www.bacongobbler.com foo.bacongobbler.com"

In confd v0.5.0, there is no way to split strings by whitespace so
there's no way to iterate through each domain bound to an application.
In this commit, there's a controller data migration script which will
migrate existing etcd keys for custom domain endpoints over to the new
syntax:

    SET /deis/domains/www.bacongobbler.com "go"
    SET /deis/domains/foo.bacongobbler.com "go"

Confd v0.8.0 has support for splitting strings, but with this approach
we can start providing support for migrations such as the one proposed
in #3399.