chore(controller): remove $PORT logic from fleet
For buildpack apps, [$PORT is exposed][1] as part of slugrunner's job.
For Dockerfile/Docker images, the internal port numbers are already
known to the developer, so if they want to follow 12-factor best
practices then they will have set up $PORT in their own image.

deis/publisher exposes the port with the lowest numerical value to the
router, which is why the documentation is reflected to show this.

https://github.com/deis/deis/blob/e6a078b4119887b4de73e8a6d04291b7cdeb823b/builder/image/slugrunner/Dockerfile#L10-L12